requirements.txt
fastapi==0.73.0
uvicorn==0.17.4
motor==3.4.0s
python-decouple==3.6